+++ CHANGELOG.md
@@ -172,6 +172,7 @@ CHANGELOG
 - Updated documentation on `skip_header` parameter (PR#2313 by Filip Pokorný).
 - Viriback Unsafe Sites feed replaced with Viriback C2 Tracker. (PR#2266 by Filip Pokorný)
 - Netlab 360 Mirai Scanner feed removed as it is discontinued. (PR#2266 by Filip Pokorný)
+- Benkow Malware Panels Tracker feed changed parser configuration. (PR#2266 by Filip Pokorný)
 
 ### Packaging
 - Remove deleted `intelmq.bots.experts.sieve.validator` from executables in `setup.py` (PR#2256 by Filip Pokorný).